# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SmartLead API",
        "description": "API para teste técnico SmartLead",
        "contact": {
            "email": "admin@smartlead.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/export/tasks/csv": {
            "get": {
                "tags": [
                    "Export"
                ],
                "summary": "Exportar tarefas para CSV",
                "operationId": "795749674f5d6bdfa07a89c470df4dc7",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrar por status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pendente",
                                "em_andamento",
                                "concluida"
                            ]
                        }
                    },
                    {
                        "name": "priority",
                        "in": "query",
                        "description": "Filtrar por prioridade",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "baixa",
                                "media",
                                "alta"
                            ]
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "description": "Data inicial",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "description": "Data final",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Arquivo CSV",
                        "content": {
                            "text/csv": {}
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/export/report/summary": {
            "get": {
                "tags": [
                    "Export"
                ],
                "summary": "Exportar relatório resumido",
                "operationId": "f574f2692246e111f2493b3f073b12ed",
                "responses": {
                    "200": {
                        "description": "Relatório resumido em JSON",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "company_name": {
                                            "type": "string"
                                        },
                                        "generated_at": {
                                            "type": "string"
                                        },
                                        "period": {
                                            "type": "string"
                                        },
                                        "summary": {
                                            "type": "object"
                                        },
                                        "tasks_by_status": {
                                            "type": "object"
                                        },
                                        "tasks_by_priority": {
                                            "type": "object"
                                        },
                                        "users_performance": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}